# backend/docs/swagger/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
...
  dto.BatchDeleteRequest:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
...
  dto.DynastyCreateRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 50
        type: string
      period:
        maxLength: 100
        type: string
    required:
    - name
    type: object
  dto.DynastyUpdateRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 50
        type: string
      period:
        maxLength: 100
        type: string
    type: object
...
  dto.ForumPostPinRequest:
    properties:
      isPinned:
        description: IsPinned 是否置顶，置顶帖列表优先展示。
        type: boolean
    type: object
...
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 识海诗词平台后端 API 文档。
  title: 识海 API
  version: "1.0"
paths:
  /api/admin/announcements:
    post:
      parameters:
      - description: 公告信息
        in: body
        name: request
        required: true
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 创建公告
      tags:
      - 后台公告
  /api/admin/announcements/{id}:
    delete:
      parameters:
      - description: 公告 ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除公告
      tags:
      - 后台公告
    put:
      parameters:
      - description: 公告 ID
        in: path
        name: id
        required: true
        type: string
      - description: 公告信息
        in: body
        name: request
        required: true
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 更新公告
      tags:
      - 后台公告
  /api/admin/comments/all:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 查询全部评论
      tags:
      - 后台评论
  /api/admin/corrections:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 查询纠错列表
      tags:
      - 后台纠错
  /api/admin/dynasties:
    delete:
      parameters:
      - description: 朝代 ID 列表
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BatchDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 批量删除朝代
      tags:
      - 后台基础数据
    post:
      parameters:
      - description: 朝代信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DynastyCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 创建朝代
      tags:
      - 后台基础数据
  /api/admin/dynasties/{id}:
    delete:
      parameters:
      - description: 朝代 ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除朝代
      tags:
      - 后台基础数据
    put:
      parameters:
      - description: 朝代 ID
        in: path
        name: id
        required: true
        type: string
      - description: 朝代信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DynastyUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 更新朝代
      tags:
      - 后台基础数据
  /api/admin/forum/posts:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 是否包含已删除帖子
        in: query
        name: includeDeleted
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 查询后台论坛帖子列表
      tags:
      - 后台论坛
  /api/admin/forum/posts/{id}:
    delete:
      parameters:
      - description: 帖子 ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 后台删除论坛帖子
      tags:
      - 后台论坛
  /api/admin/forum/posts/{id}/pin:
    put:
      parameters:
      - description: 帖子 ID
        in: path
        name: id
        required: true
        type: string
      - description: 置顶状态
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForumPostPinRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 设置论坛帖子置顶状态
      tags:
      - 后台论坛
  /api/admin/forum/replies/{id}:
    delete:
      parameters:
      - description: 回复 ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 后台删除论坛回复
      tags:
      - 后台论坛
  /api/admin/poem-annotations/{id}:
    delete:
      parameters:
      - description: 标注 ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除诗词标注
      tags:
      - 后台诗词标注
    put:
      parameters:
      - description: 标注 ID
        in: path
        name: id
        required: true
        type: string
      - description: 标注信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PoemAnnotationUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 更新诗词标注
      tags:
      - 后台诗词标注
  /api/admin/poems:
    delete:
      parameters:
      - description: 诗词 ID 列表
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BatchDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 批量删除诗词
      tags:
      - 后台诗词
    post:
      parameters:
      - description: 诗词信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PoemCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 创建诗词
      tags:
      - 后台诗词
  /api/admin/poems/{id}:
    delete:
      parameters:
      - description: 诗词 ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除诗词
      tags:
      - 后台诗词
    put:
      parameters:
      - description: 诗词 ID
        in: path
        name: id
        required: true
        type: string
      - description: 诗词信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PoemUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 更新诗词
      tags:
      - 后台诗词
...
  /api/announcements/{id}:
    get:
      parameters:
      - description: 公告 ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取公告详情
      tags:
      - 公告
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"